requirements.txt
Django==3.1.6
django-crispy-forms==1.11.1
django-compressor==2.4
djangorestframework==3.12.2
pytest==6.2.2
pytest-django==4.1.0
pytest-factoryboy==2.1.0
python-dotenv==0.15.0
selenium==3.141.0
Pillow==8.1.0